--- app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/ce/PingScheduledTaskCE.java
@@ -1,6 +1,0 @@
-package com.appsmith.server.solutions.ce;
-
-public interface PingScheduledTaskCE {
-
-    void pingSchedule();
-}
