+++ app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/ce/ScheduledTaskCE.java
@@ -1,5 +1,6 @@
 package com.appsmith.server.solutions.ce;
 
 public interface ScheduledTaskCE {
-    void fetchFeatures();
+
+    void pingSchedule();
 }
